build(app-shell): Switch to babel compiled source
This change allows the app-shell project to start using flow and also import the newly written (and
babel-dependent) DiscoveryClient
