+++ app-shell/.gitignore
@@ -1,0 +1,2 @@
+# lib is compiled by babel
+lib
