+++ package.json
@@ -56,6 +56,7 @@
   },
   "devDependencies": {
     "ajv": "^6.2.1",
+    "babel-cli": "^6.26.0",
     "babel-core": "^6.25.0",
     "babel-eslint": "7.2.3",
     "babel-loader": "^7.1.1",
@@ -94,6 +95,7 @@
     "jest": "^20.0.4",
     "lerna": "^3.0.0-beta.21",
     "lost": "^8.2.0",
+    "nodemon": "^1.18.3",
     "portfinder": "^1.0.13",
     "postcss": "^6.0.16",
     "postcss-cssnext": "^3.0.2",
